+++ src/components/my-finances/voucher-create-form.tsx
@@ -37,10 +37,7 @@ const VoucherCreateForm = memo(function VoucherCreateForm({
   );
 
   return (
-    <form
-      onSubmit={onSubmit}
-      className="font-custom flex flex-col gap-4 pt-2"
-    >
+    <form onSubmit={onSubmit} className="font-custom flex flex-col gap-4 pt-2">
       <div className="space-y-2">
         <Label htmlFor="voucher-date">Date</Label>
         <Input